requirements.txt
# Core Web Framework
streamlit==1.28.1

# NLP and Machine Learning
transformers==4.36.2
torch==2.1.2
nltk==3.8.1

# Data Processing
pandas==2.1.3
numpy==1.24.3

# Visualization
plotly==5.18.0
matplotlib==3.8.2

# Utilities
python-dotenv==1.0.0
